+++ tezos/client/src/client.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 use crypto::hash::{ChainId, ContextHash, ProtocolHash};
-use tezos_api::ffi::{ApplyBlockError, ApplyBlockRequest, ApplyBlockResponse, BeginConstructionError, BeginConstructionRequest, CommitGenesisResult, ContextDataError, GenesisChain, GetDataError, InitProtocolContextResult, JsonRpcResponse, PatchContext, PrevalidatorWrapper, ProtocolJsonRpcRequest, ProtocolOverrides, ProtocolRpcError, TezosGenerateIdentityError, TezosRuntimeConfiguration, TezosRuntimeConfigurationError, TezosStorageInitError, ValidateOperationError, ValidateOperationRequest, ValidateOperationResponse};
+use tezos_api::ffi::{ApplyBlockError, ApplyBlockRequest, ApplyBlockResponse, BeginConstructionError, BeginConstructionRequest, CommitGenesisResult, ContextDataError, GenesisChain, GetDataError, InitProtocolContextResult, JsonRpcResponse, PatchContext, PrevalidatorWrapper, ProtocolJsonRpcRequest, ProtocolOverrides, ProtocolRpcError, TezosGenerateIdentityError, TezosRuntimeConfiguration, TezosRuntimeConfigurationError, TezosStorageInitError, ValidateOperationError, ValidateOperationRequest, ValidateOperationResponse, ComputePathError, ComputePathRequest, ComputePathResponse};
 use tezos_api::identity::Identity;
 use tezos_interop::{ffi, runtime};
 
@@ -115,6 +115,20 @@ pub fn call_protocol_json_rpc(request: ProtocolJsonRpcRequest) -> Result<JsonRpc
         }
     }
 }
+
+/// Call compute path
+/// TODO: TE-207 Implement in Rust
+pub fn compute_path(request: ComputePathRequest) -> Result<ComputePathResponse, ComputePathError> {
+    match ffi::compute_path(request) {
+        Ok(result) => Ok(result?),
+        Err(e) => {
+            Err(ComputePathError::PathError {
+                message: format!("Path computation failed! Reason: {:?}", e)
+            })
+        }
+    }
+}
+
 
 /// Call helpers_preapply_operations shell service
 pub fn helpers_preapply_operations(request: ProtocolJsonRpcRequest) -> Result<JsonRpcResponse, ProtocolRpcError> {